+++ rollup.config.mjs
@@ -1,6 +1,10 @@
 
-const config =  {
+const config = {
   input: "build/es6/antity.js",
+  // onwarn(warning, warn) {
+  //   if (warning.code === "THIS_IS_UNDEFINED") return;
+  //   warn(warning);
+  // },
   output: {
     name: "antity",
     file: "build/antity.mjs",