+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 
+# 0.17.0 (May 9th 2026)
+
+- Update @dwtechs/checkard to 3.6.0
+- Update @dwtechs/winstan to 0.7.0
+- Add `ansiEscapeCode`, `locale` and `timeZone` types
+
 # 0.16.0 (Mar 8th 2026)
 
 - Rename `send` property to `isPrivate` with inverted logic (isPrivate: true means the property should NOT be sent)